refactor(misc1): move way beep functions elsewhere
