+++ src/nvim/move.c
@@ -23,6 +23,7 @@
 #include "nvim/diff.h"
 #include "nvim/edit.h"
 #include "nvim/fold.h"
+#include "nvim/getchar.h"
 #include "nvim/mbyte.h"
 #include "nvim/memline.h"
 #include "nvim/misc1.h"
